Defined a_star()
Defined a_star() function which takes in a graph containing edges between nodes and using a priorityQueue finds the shortest path between start and end point in the graph. heuristics() is needed to compute the priority of the current path: if the move takes us further from the end point, it will not be a part of the optimal path. Also changed coordinate variables to be integers in roomGeneration.py and listMatrix.py