Defined whortest_path and build_path functions
shortest_path takes in the result of a_star() and creates a list containing the points belonging to the shortest path from point a to point b. The build_path function uses this list of the shortest path and inputs the paths to a matrix which will then contain the necessary information for drawing tiles on the screen.